+++ .gitignore
@@ -3,3 +3,4 @@ node_modules
 dist/
 yarn.lock
 .DS_Store
+.npmrc
