+++ server/db.js
@@ -1,0 +1,6 @@
+const Sequelize = require("sequelize");
+const db = new Sequelize("postgres://localhost:5432/SEP", {
+  logging: false
+});
+
+module.exports = { db };
