+++ server/models/campuses.js
@@ -1,4 +1,4 @@
-const { db } = require("../server/index");
+const { db } = require("../db");
 const Sequelize = require("sequelize");
 const Campus = db.define("campus", {
   name: {
