+++ server/models/students.js
@@ -1,5 +1,5 @@
 const Sequelize = require("sequelize");
-const { db } = require("../server/index");
+const { db } = require("../db");
 const Student = db.define("student", {
   firstName: {
     type: Sequelize.STRING,
